chore: rename "logic" directory to "utils" for better clarity
