--- src/logic/index.ts
@@ -1,5 +1,0 @@
-export * from "./deckLogic";
-export * from "./fetchWithAuth";
-export * from "./statsLogic";
-export * from "./tableLogic";
-export * from "./userUtils";
